feat: Rewrite Client Log parser
